fix(cli): keep exec.rs within the file-size budget and build under --all-features
- Lift --host-service validation into its own commands/vm/host_services module,
  which also collapses a duplicated sidecar resolution: the bindings and the
  attachment now come from one call, so the plan grant and the attached volume
  cannot disagree about which bytes are mounted.
- The mcp-gated ExecRequest literal needed the new field; the local check had
  not been run with --all-features.
